+++ forui/lib/src/widgets/date_field/field/date_field.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:forui/src/foundation/field/field.dart';
 import 'package:forui/src/foundation/field/field_controller.dart';
+import 'package:forui/src/localizations/localization.dart';
 
 import 'package:meta/meta.dart';
 
@@ -82,7 +83,12 @@ class _DateFieldState extends FieldState<DateField, DateTime?> {
   @override
   void didUpdateWidget(covariant DateField old) {
     super.didUpdateWidget(old);
-    if (widget.calendarController != old.calendarController) {
+    if (widget.localizations != old.localizations) {
+      localizations = scriptNumerals.contains(widget.localizations.localeName) ? FDefaultLocalizations() : widget.localizations;
+      controller.dispose();
+      controller = createController();
+
+    } else if (widget.calendarController != old.calendarController) {
       controller.dispose();
       controller = createController();
     }