+++ forui/lib/src/widgets/date_field/field/date_field_controller.dart
@@ -111,6 +111,12 @@ class DateFieldController extends FieldController {
       );
     }
   }
+
+  @override
+  void dispose() {
+    controller.removeListener(updateFromCalendar);
+    super.dispose();
+  }
 }
 
 @internal